Fix issue where devices get stuck at the upper end of the bounding box
If the initial guess for the Nelder-Mead algorithm is close to the upper end of
the bounding box Math.Net will sometimes fail to pick any test points that are
inside the bounding box. It then falls back to using the initial guess as the
solution, causing devices to get stuck at that location until the app is
restarted. This seems to happen because the error value for those test points
is all indentical (i.e. positive infinity).

This patch changes the error value calculation to add a gradient based on how
far a test point is outside of the bounding box. This gives the optimization
algorithm a sense of directionality. It also clamps the minimizing point to
the bounding box in case we do end up with a solution that's outside of the
bounding box.

This problem can be reproduced by setting at least two components of the the
initial guess vector to the corresponding values of the upper bound of the
bounding box:

For example, with bounds: [[0, 0, 0], [15.26, 7.78, 2.5]]:

guess = new Point3D(15.26, 3.0, 2.5); // causes devices to get stuck
guess = new Point3D(15.25, 3.0, 2.5); // does not cause devices to get stuck