Fix issue where devices can get stuck at X=0, Y=0 and Z=0
This can happen because the initial test point picked by Math.Net is only
0.00025 units (for each vector component that is 0) away from the initial
guess.

With this patch we're picking a (hopefully) better starting point by projecting
a unit vector from the initial guess towards the centroid of the triangle
that's described by the locations of the closest three nodes.

This problem can be reproduced by setting all vector components of the initial
guess to 0:

guess = new Point3D(0, 0, 0);

Most of the time some of the vector components of the solution will be zero
temporarily - and at least some of the time the solution will converge on a
value where at least one vector component is permanently zero until the app is
restarted.